Change NL_LOG_CREATE_ENTRY() to generate a static nl_log_token_region_entry_t
For the entry that goes into the .token_region_table, which
should be a nl_log_token_entry_t, the NL_LOG_CREATE_ENTRY() macro
used to generate two static uint32_t values, and rely on the
compiler/linker to order them correctly when collected into
the .token_region_table. This is unreliable and we've seen
it break because the order changed with newer version of gcc.

Fix NL_LOG_CREAT_ENTRY() to create a nl_log_token_entry_t
as it should have in the first place. However, we have to
fix other places which define macros which might take NL_LOG()
as an argument like nlassert's nlREQUIRE_ACTION(), because
the extra comma in the static initializer for the struct can
cause the preprocessor to falsely report that we're providing
too many arguments to the outer macro.

The workaround is to make the outer macro be variadic, even
if it doesn't really take variable number of arguments. This
avoids the false alarm warning/error, though we do risk losing
other errors when someone does really pass the wrong number of
arguments to the macro.