Add symbol openocd needs for FreeRTOS task awareness
FreeRTOS used to have the symbol uxTopUsedPriority, which openocd
used to determine the number of elements in the the pxReadyTasksLists
array. A similar variable, uxTopReadyPriority, was added to later
versions of FreeRTOS, which we extended openocd to use if it didn't
find uxTopUsedPriority, but using uxTopReadyPriority is problematic
because it's value changes depending on the value of
configUSE_PORT_OPTIMISED_TASK_SELECTION in FreeRTOSConfig.h.

If that config is 0, the value is simple uint, which is equivalent
to the uxTopReadyPriority once all tasks have been created (i.e.
it's not a constant). If that config is 1, the value may be
equal to "1 << highest ready priority", depending on the port.
Openocd has no way to tell which representation is being used.

Best thing is to just define uxTopUsedPriority ourselves so
there is no issue with compiler setting.

Products need to make sure the symbol isn't dead stripped by
adding an EXTERN(uxTopUsedPriority) to their app linker script.