check the value of NLER_FEATURE_SIMULATEABLE_TIME
NLER was inconsistent about its treatment of the flag. Looks
like the idea was that NLER_FEATURE_* flags are always defined,
but sometimes they're zeroed.

Also, we consider NLER_FEATURE_SIMULATEABLE_TIME and
NLER_FEATURE_TIMER_USING_SWTIMER to be mutually exclusive,
so it's redundant to check them both - in fact, that was why
the error has not surfaced before.